Rename defaults/ to default, fix some comments.
