+++ src/default/plugins.js
